consolodate query and types
