+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r", encoding="utf-8") as fh:
 
 setuptools.setup(
     name="duneapi",
-    version="2.0.0",
+    version="2.0.2",
     author="Benjamin H. Smith",
     author_email="bh2smith@gmail.com",
     description="A simple framework for interacting with Dune Analytics' unsupported API.",
@@ -23,5 +23,8 @@ setuptools.setup(
     install_requires=["requests>=2.27.1", "types-requests>=2.27.13"],
     package_dir={"": "src"},
     packages=setuptools.find_packages(where="src"),
+    package_data={
+        "duneapi": ["py.typed"],
+    },
     python_requires=">=3.9",
 )
