+++ src/duneapi/api.py
@@ -12,12 +12,11 @@ from typing import Optional
 
 from requests import Session, Response
 
-from src.duneapi.query import DuneQuery, Post
 from src.duneapi.response import (
     validate_and_parse_dict_response,
     validate_and_parse_list_response,
 )
-from src.duneapi.types import DuneRecord, QueryResults
+from src.duneapi.types import DuneRecord, QueryResults, DuneQuery, Post
 
 log = logging.getLogger(__name__)
 logging.config.fileConfig(fname="logging.conf", disable_existing_loggers=True)
