+++ src/duneapi/response.py
@@ -3,8 +3,7 @@ from typing import Any
 
 from requests import Response
 
-from src.duneapi.query import KeyMap
-from src.duneapi.types import ListInnerResponse, DictInnerResponse
+from src.duneapi.types import ListInnerResponse, DictInnerResponse, KeyMap
 
 
 def pre_validate_response(response: Response, key_map: KeyMap) -> dict[str, Any]:
